Update status_code to uint16_t and add HTTP response tests
Change status_code type from uint8_t to uint16_t for better range handling. Add tests for rendering HTTP 404 response with no body and 200 response with a JSON body to ensure proper functionality.

+++ src/tiny_http/tiny_http_server_lib.h
@@ -39,7 +39,7 @@ typedef struct http_request {
 
 typedef struct http_response {
     http_version version;
-    uint8_t status_code;
+    uint16_t status_code;
     uint8_t *reason_phrase;
     http_header *headers;
     size_t headers_cnt;
